PRSD-451: Moves email models into their own package
