+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/models/emailModels/ExampleEmail.kt
@@ -1,4 +1,4 @@
-package uk.gov.communities.prsdb.webapp.models.viewModels
+package uk.gov.communities.prsdb.webapp.models.emailModels
 
 data class ExampleEmail(
     var firstName: String,
